天使令牌相关接口

+++ src/main/java/com/lhjl/tzzs/proxy/mapper/UserIntegralsMapper.java
@@ -1,6 +1,6 @@
 package com.lhjl.tzzs.proxy.mapper;
 
-import java.util.List;
+import java.math.BigDecimal;
 import java.util.Map;
 
 import org.apache.ibatis.annotations.Mapper;
@@ -11,8 +11,8 @@ import com.lhjl.tzzs.proxy.utils.OwnerMapper;
 @Mapper
 public interface UserIntegralsMapper extends OwnerMapper<UserIntegrals> {
 
-	Integer findIntegralsZ(@Param("userId") Integer userId);
-	Integer findIntegralsX(@Param("userId") Integer userId);
+	BigDecimal findIntegralsZ(@Param("userId") Integer userId);
+	BigDecimal findIntegralsX(@Param("userId") Integer userId);
 	String findBySkey(@Param("leId") Integer leId);
 	Integer findByQnum (@Param("leId") Integer leId);
 	Map<String,Object>findIntegralsU(@Param("userId") Integer userId);
